refactor(mcp): align tool contracts with surface owners
