+++ crates/coral-mcp/src/surface/arguments.rs
@@ -1,0 +1,75 @@
+use rmcp::ErrorData;
+use serde_json::{Map, Value};
+
+pub(crate) fn required_string_argument(
+    arguments: Option<&Map<String, Value>>,
+    key: &str,
+) -> Result<String, ErrorData> {
+    let value = arguments
+        .and_then(|arguments| arguments.get(key))
+        .and_then(Value::as_str)
+        .map(str::trim)
+        .filter(|value| !value.is_empty())
+        .ok_or_else(|| {
+            ErrorData::invalid_params(format!("missing string argument '{key}'"), None)
+        })?;
+    Ok(value.to_string())
+}
+
+pub(crate) fn optional_string_argument(
+    arguments: Option<&Map<String, Value>>,
+    key: &str,
+) -> Result<Option<String>, ErrorData> {
+    let Some(value) = arguments.and_then(|arguments| arguments.get(key)) else {
+        return Ok(None);
+    };
+    if value.is_null() {
+        return Ok(None);
+    }
+    let value = value.as_str().ok_or_else(|| {
+        ErrorData::invalid_params(format!("argument '{key}' must be a string"), None)
+    })?;
+    let value = value.trim();
+    if value.is_empty() {
+        Ok(None)
+    } else {
+        Ok(Some(value.to_string()))
+    }
+}
+
+pub(crate) fn optional_non_empty_string_argument(
+    arguments: Option<&Map<String, Value>>,
+    key: &str,
+) -> Result<Option<String>, ErrorData> {
+    let Some(value) = arguments.and_then(|arguments| arguments.get(key)) else {
+        return Ok(None);
+    };
+    if value.is_null() {
+        return Ok(None);
+    }
+    let value = value.as_str().ok_or_else(|| {
+        ErrorData::invalid_params(format!("argument '{key}' must be a string"), None)
+    })?;
+    let value = value.trim();
+    if value.is_empty() {
+        Err(ErrorData::invalid_params(
+            format!("argument '{key}' must not be empty"),
+            None,
+        ))
+    } else {
+        Ok(Some(value.to_string()))
+    }
+}
+
+pub(crate) fn optional_bool_argument(
+    arguments: Option<&Map<String, Value>>,
+    key: &str,
+    default: bool,
+) -> Result<bool, ErrorData> {
+    let Some(value) = arguments.and_then(|arguments| arguments.get(key)) else {
+        return Ok(default);
+    };
+    value.as_bool().ok_or_else(|| {
+        ErrorData::invalid_params(format!("argument '{key}' must be a boolean"), None)
+    })
+}
